+++ ohlcv-ctl/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [0.0.2](https://github.com/typedduck/ohlcv/tree/ohlcv-ctl-v0.0.2) - 2024-08-27
+
+- Updated dependencies to `ohlc` v0.0.2.
+
 ## [0.0.1](https://github.com/typedduck/ohlcv/tree/ohlcv-ctl-v0.0.1) - 2024-08-27
 
 - Initial release